Create Random_Password_Generator
Project 3: password generator using python. This code can be made to advanced mode and link it with React to build and app just like Authenticator instead of numbers only, can be used as a password generator. Hook it to a database and add as a extension, the voila, you have got a unique password generator for all your app to be secure and safe where no one can crack it.